Remove addon-related methods from cloud botanist interface
See also: https://github.com/gardener/gardener-extensions/pull/235
