+++ charts/shoot-core/charts/vpn-shoot/templates/deployment.yaml
@@ -34,24 +34,6 @@ spec:
       tolerations:
       - key: CriticalAddonsOnly
         operator: Exists
-{{- if .Values.initContainers }}
-      initContainers:
-      {{- range $index, $ctr := .Values.initContainers }}
-      - name: init-{{ $index }}
-        image: {{ $ctr.image }}
-        {{- if $ctr.command }}
-        command:
-{{ toYaml $ctr.command | indent 8 }}
-        {{- end }}
-        {{- if $ctr.args }}
-        args:
-{{ toYaml $ctr.args | indent 8 }}
-        {{- end }}
-        volumeMounts:
-        - name: shared-init-config
-          mountPath: /init-config
-      {{- end }}
-{{- end }}
       containers:
       - name: vpn-shoot
         image: {{ index .Values.images "vpn-shoot" }}
@@ -82,8 +64,6 @@ spec:
           name: vpn-shoot-tlsauth
         - mountPath: /srv/secrets/dh
           name: vpn-shoot-dh
-        - name: shared-init-config
-          mountPath: /init-config
       volumes:
       - name: vpn-shoot
         secret:
@@ -94,5 +74,3 @@ spec:
       - name: vpn-shoot-dh
         secret:
           secretName: vpn-shoot-dh
-      - name: shared-init-config
-        emptyDir: {}
